+++ midas/midas_net.py
@@ -27,7 +27,7 @@ class MidasNet(BaseModel):
 
         use_pretrained = False if path is None else True
 
-        self.pretrained, self.scratch = _make_encoder(features, use_pretrained)
+        self.pretrained, self.scratch = _make_encoder(backbone="resnext101_wsl", features=features, use_pretrained=use_pretrained)
 
         self.scratch.refinenet4 = FeatureFusionBlock(features)
         self.scratch.refinenet3 = FeatureFusionBlock(features)
