+++ ros/additions/downloads.sh
@@ -1,5 +1,5 @@
 mkdir ~/.ros
-wget https://github.com/intel-isl/MiDaS/releases/download/v2/model_opt.pt
-cp ./model_opt.pt ~/.ros/model_opt.pt
+wget https://github.com/intel-isl/MiDaS/releases/download/v2_1/model-small-traced.pt
+cp ./model-small-traced.pt ~/.ros/model-small-traced.pt
