+++ ros/launch_midas_cpp.sh
@@ -1,2 +1,2 @@
 source ~/catkin_ws/devel/setup.bash
-roslaunch midas_cpp midas_cpp.launch model_name:="model_opt.pt" input_topic:="image_topic" output_topic:="midas_topic" out_orig_size:="true"
+roslaunch midas_cpp midas_cpp.launch model_name:="model-small-traced.pt" input_topic:="image_topic" output_topic:="midas_topic" out_orig_size:="true"
